入力を整数化

+++ backend/static/problems/perceptron/solution.py
@@ -2,9 +2,9 @@
 import numpy as np
 
 # 入力の受け取り
-x = np.array(list(map(float, input().split())))
-w = np.array(list(map(float, input().split())))
-b = float(input())
+x = np.array(list(map(int, input().split())))
+w = np.array(list(map(int, input().split())))
+b = int(input())
 
 # 計算
 y = np.dot(w, x) + b
